+++ app/src/main/java/de/hamedtanha/servertoolkit/core/di/DatabaseModule.kt
@@ -9,6 +9,7 @@ import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
 import de.hamedtanha.servertoolkit.core.database.MIGRATION_1_2
 import de.hamedtanha.servertoolkit.core.database.MIGRATION_2_3
+import de.hamedtanha.servertoolkit.core.database.MIGRATION_3_4
 import de.hamedtanha.servertoolkit.core.database.ServerToolkitDatabase
 import javax.inject.Singleton
 
@@ -26,7 +27,7 @@ object DatabaseModule {
             ServerToolkitDatabase::class.java,
             DATABASE_NAME,
         )
-            .addMigrations(MIGRATION_1_2, MIGRATION_2_3)
+            .addMigrations(MIGRATION_1_2, MIGRATION_2_3, MIGRATION_3_4)
             .build()
     }